pinctrl: qcom: Clear output enable bit for GPIO direct connect
On wahoo, the touchscreens use direct connect and cause an IRQ storm on
boot (or indefinitely in recovery mode). The IRQ storm results in thousands
of interrupts being sent per second, which becomes a significant source of
overhead.

Setting the direct-connect GPIOs to input mode (i.e., clearing the output
enable bit) fixes the IRQ storm.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>