simple_lmk: Introduce simple low memory killer for Android
This is a simple replacement for the Android Low Memory Killer driver that
simply targets a predefined number of pages to be freed every time a memory
allocation attempt is about to utilize direct reclaim. Processes are killed
according to the priorities that Android gives them; in other words, the
least important processes are always killed first.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>